数据库字段类型映射

+++ framework-codegen/src/main/java/org/unicorn/framework/codegen/mapper/MysqlDbCloumnTypeMapper.java
@@ -17,7 +17,7 @@ public class MysqlDbCloumnTypeMapper {
 		_MAP.put("TEXT", new DbCloumnTypeInfo("TEXT", null,"java.lang.String"));
 		_MAP.put("LONGTEXT", new DbCloumnTypeInfo("TEXT", null,"java.lang.String"));
 		_MAP.put("TINYINT", new DbCloumnTypeInfo("INTEGER", null,"java.lang.Integer"));
-		_MAP.put("SMALLINT", new DbCloumnTypeInfo("TINYINT", null,"java.lang.Integer"));
+		_MAP.put("SMALLINT", new DbCloumnTypeInfo("INTEGER", null,"java.lang.Integer"));
 		_MAP.put("INT", new DbCloumnTypeInfo("INTEGER", null,"java.lang.Integer"));
 		_MAP.put("INTEGER", new DbCloumnTypeInfo("INTEGER", null,"java.lang.Integer"));
 		_MAP.put("BIGINT", new DbCloumnTypeInfo("BIGINT", null,"java.lang.Long"));
